Housing Data
This is all the housing data, not restricted to Suffolk county yet. There are 2 rows of headings (non english and english) so you will have to mindful of that when reading from it (or just open it in excel and modify it). I am not doing anything with it yet because we have plenty of other data to focus on.